chore: Initial chop up of slider
