+++ projects/Mallard/src/screens/article/slider/types.ts
@@ -1,0 +1,6 @@
+export interface SliderSection {
+    items: number
+    title: string
+    color: string
+    startIndex: number
+}
